Shells/cloudshell-L1-l1mock/dependencies/cloudshell-l1-networking-core/requirements.txt
cloudshell-core==2.1.153

